# cpu/cc2538/periph/timer: handle power modes