### Contribution description

This PR adds enable gating `GPT` clock in `sleep` and `PM0`.

When testing this PR I had an issue with `tests/periph_timer`. The api for `timer` says that after `timer_stop` the underlying peripheral should be completely powered off.

Doing this for `cc2538` makes writing to some registers not possible, namely `TxMATCHR ` written to in `timer_set_absolute`. This makes `tests/periph_timer` and others where `timer_stop` is called fail.

To make the tests pass c0034a0d72ec5d1b897d914fdc06f0e2a9b408bb enables the clock gate while configuring the timer. But this adds more handling in `timer_set_absolulte` which increases the offset in `tests/xtimer_usleep` in `3us`.

### Testing procedure

- `BOARD=openmote-b make -C tests/xtimer_periodic_wakeup/ flash test` passes (fails in master)
- Compare offset in master and this PR for `BOARD=openmote-b make -C tests/xtimer_usleep flash test`
- `BOARD=openmote-b make -C tests/periph_timer flash test` works the same as in master (two timer configuration fail because `In 32-bit mode, the GPTimer frequency must equal the system clock frequency (%u)`).

### Issues/PRs references

Alternative to #13645