@fjmolinas 2020-03-19
> But I get the same results in `tests/xtimer_usleep`.

Well I think the increase is something we will just have to accept for now for the `timer` implementation to be api compliant.