@fjmolinas 2020-03-19
> I though what if instead of enabling the timer, setting the timer and disabling it again, we store the set value and apply it the next time the timer would be enabled:

What are the pro and cons? I dont really know how to make the decision hahaha :)