@benpicco 2020-03-19
I thought enabling / disabling the timer would cause a greater overhead, but I didn't measure it.
If it's negligible, it's the preferable solution since it's less complex and doesn't require additional memory.