@fjmolinas 2020-03-19
> I thought enabling / disabling the timer would cause a greater overhead, but I didn't measure it.
> If it's negligible, it's the preferable solution since it's less complex and doesn't require additional memory.

I tested it a little more and went with your proposal, although it adds some code enabling and disabling the timier when calling `set_absolute` made the first timer elapse a little early.