@fjmolinas 2020-03-19
I'll try to adapt the test for `periph_timer` so it works for this platform as well and I will post comparative results.