@benpicco 2020-03-21
If you are happy with the current state of this PR, squash.
This fixes a serious regression on cc2538.