@fjmolinas 2020-03-23
Some comparative test restuls

**run `tests/periph_timer`:**

- master

<details><summary>BOARD=openmote-b make -C tests/periph_timer flash -j3</summary>

```
2020-03-23 09:54:59,931 # main(): This is RIOT! (Version: 2020.04-devel-1418-g5ef0c-HEAD)
2020-03-23 09:54:59,932 # 
2020-03-23 09:54:59,944 # Test for peripheral TIMERs
2020-03-23 09:54:59,945 # 
2020-03-23 09:54:59,945 # Available timers: 4
2020-03-23 09:54:59,945 # 
2020-03-23 09:54:59,946 # Testing TIMER_0:
2020-03-23 09:54:59,946 # TIMER_0: initialization successful
2020-03-23 09:54:59,947 # TIMER_0: stopped
2020-03-23 09:54:59,947 # TIMER_0: set channel 0 to 5000
2020-03-23 09:54:59,948 # TIMER_0: set channel 1 to 10000
2020-03-23 09:54:59,948 # TIMER_0: starting
2020-03-23 09:54:59,961 # TIMER_0: channel 0 fired at SW count      312 - init:      312
2020-03-23 09:54:59,963 # TIMER_0: channel 1 fired at SW count      620 - diff:      308
2020-03-23 09:54:59,963 # 
2020-03-23 09:54:59,964 # Testing TIMER_1:
2020-03-23 09:54:59,964 # TIMER_1: initialization successful
2020-03-23 09:54:59,976 # TIMER_1: stopped
2020-03-23 09:54:59,977 # TIMER_1: set channel 0 to 5000
2020-03-23 09:54:59,977 # TIMER_1: starting
2020-03-23 09:54:59,978 # TIMER_1: channel 0 fired at SW count      313 - init:      313
2020-03-23 09:54:59,978 # 
2020-03-23 09:54:59,979 # Testing TIMER_2:
2020-03-23 09:54:59,980 # TIMER_2: initialization successful
2020-03-23 09:54:59,992 # TIMER_2: stopped
2020-03-23 09:54:59,993 # TIMER_2: set channel 0 to 5000
2020-03-23 09:54:59,994 # TIMER_2: set channel 1 to 10000
2020-03-23 09:54:59,995 # TIMER_2: starting
2020-03-23 09:54:59,996 # TIMER_2: channel 0 fired at SW count      312 - init:      312
2020-03-23 09:55:00,009 # TIMER_2: channel 1 fired at SW count      620 - diff:      308
2020-03-23 09:55:00,009 # 
2020-03-23 09:55:00,010 # Testing TIMER_3:
2020-03-23 09:55:00,011 # TIMER_3: initialization successful
2020-03-23 09:55:00,012 # TIMER_3: stopped
2020-03-23 09:55:00,013 # TIMER_3: set channel 0 to 5000
2020-03-23 09:55:00,014 # TIMER_3: starting
2020-03-23 09:55:00,025 # TIMER_3: channel 0 fired at SW count      313 - init:      313
2020-03-23 09:55:00,025 # 
2020-03-23 09:55:00,025 # TEST SUCCEEDED
```
</details>

- pr

<details><summary>BOARD=openmote-b make -C tests/periph_timer flash -j3</summary>

```
2020-03-23 09:54:39,692 # main(): This is RIOT! (Version: 2020.04-devel-1423-gb74219-pr_cc2538_timer_pm)
2020-03-23 09:54:39,692 # 
2020-03-23 09:54:39,693 # Test for peripheral TIMERs
2020-03-23 09:54:39,693 # 
2020-03-23 09:54:39,694 # Available timers: 4
2020-03-23 09:54:39,694 # 
2020-03-23 09:54:39,706 # Testing TIMER_0:
2020-03-23 09:54:39,706 # TIMER_0: initialization successful
2020-03-23 09:54:39,706 # TIMER_0: stopped
2020-03-23 09:54:39,707 # TIMER_0: set channel 0 to 5000
2020-03-23 09:54:39,707 # TIMER_0: set channel 1 to 10000
2020-03-23 09:54:39,708 # TIMER_0: starting
2020-03-23 09:54:39,722 # TIMER_0: channel 0 fired at SW count      279 - init:      279
2020-03-23 09:54:39,722 # TIMER_0: channel 1 fired at SW count      585 - diff:      306
2020-03-23 09:54:39,722 # 
2020-03-23 09:54:39,723 # Testing TIMER_1:
2020-03-23 09:54:39,724 # TIMER_1: initialization successful
2020-03-23 09:54:39,724 # TIMER_1: stopped
2020-03-23 09:54:39,737 # TIMER_1: set channel 0 to 5000
2020-03-23 09:54:39,738 # TIMER_1: starting
2020-03-23 09:54:39,739 # TIMER_1: channel 0 fired at SW count      293 - init:      293
2020-03-23 09:54:39,739 # 
2020-03-23 09:54:39,740 # Testing TIMER_2:
2020-03-23 09:54:39,741 # TIMER_2: initialization successful
2020-03-23 09:54:39,741 # TIMER_2: stopped
2020-03-23 09:54:39,742 # TIMER_2: set channel 0 to 5000
2020-03-23 09:54:39,754 # TIMER_2: set channel 1 to 10000
2020-03-23 09:54:39,754 # TIMER_2: starting
2020-03-23 09:54:39,755 # TIMER_2: channel 0 fired at SW count      279 - init:      279
2020-03-23 09:54:39,755 # TIMER_2: channel 1 fired at SW count      585 - diff:      306
2020-03-23 09:54:39,755 # 
2020-03-23 09:54:39,770 # Testing TIMER_3:
2020-03-23 09:54:39,771 # TIMER_3: initialization successful
2020-03-23 09:54:39,771 # TIMER_3: stopped
2020-03-23 09:54:39,772 # TIMER_3: set channel 0 to 5000
2020-03-23 09:54:39,773 # TIMER_3: starting
2020-03-23 09:54:39,774 # TIMER_3: channel 0 fired at SW count      293 - init:      293
2020-03-23 09:54:39,775 # 
2020-03-23 09:54:39,785 # TEST SUCCEEDED

```
</details>

**run `tests/xtimer_usleep`**

- master

<details><summary>BOARD=openmote-b make -C tests/xtimer_usleep flash -j3<b> offset: 57us</b></summary>

```
2020-03-23 09:54:12,060 # Running test 5 times with 7 distinct sleep times
2020-03-23 09:54:12,075 # Slept for 10057 us (expected: 10000 us) Offset: 57 us
2020-03-23 09:54:12,124 # Slept for 50057 us (expected: 50000 us) Offset: 57 us
2020-03-23 09:54:12,140 # Slept for 10291 us (expected: 10234 us) Offset: 57 us
2020-03-23 09:54:12,203 # Slept for 56837 us (expected: 56780 us) Offset: 57 us
2020-03-23 09:54:12,219 # Slept for 12179 us (expected: 12122 us) Offset: 57 us
2020-03-23 09:54:12,316 # Slept for 98822 us (expected: 98765 us) Offset: 57 us
2020-03-23 09:54:12,397 # Slept for 75057 us (expected: 75000 us) Offset: 57 us
2020-03-23 09:54:12,413 # Slept for 10057 us (expected: 10000 us) Offset: 57 us
2020-03-23 09:54:12,460 # Slept for 50057 us (expected: 50000 us) Offset: 57 us
2020-03-23 09:54:12,476 # Slept for 10291 us (expected: 10234 us) Offset: 57 us
2020-03-23 09:54:12,540 # Slept for 56837 us (expected: 56780 us) Offset: 57 us
2020-03-23 09:54:12,556 # Slept for 12179 us (expected: 12122 us) Offset: 57 us
2020-03-23 09:54:12,651 # Slept for 98822 us (expected: 98765 us) Offset: 57 us
2020-03-23 09:54:12,732 # Slept for 75057 us (expected: 75000 us) Offset: 57 us
2020-03-23 09:54:12,748 # Slept for 10057 us (expected: 10000 us) Offset: 57 us
2020-03-23 09:54:12,796 # Slept for 50057 us (expected: 50000 us) Offset: 57 us
2020-03-23 09:54:12,812 # Slept for 10291 us (expected: 10234 us) Offset: 57 us
2020-03-23 09:54:12,876 # Slept for 56837 us (expected: 56780 us) Offset: 57 us
2020-03-23 09:54:12,892 # Slept for 12179 us (expected: 12122 us) Offset: 57 us
2020-03-23 09:54:12,989 # Slept for 98822 us (expected: 98765 us) Offset: 57 us
2020-03-23 09:54:13,068 # Slept for 75057 us (expected: 75000 us) Offset: 57 us
2020-03-23 09:54:13,085 # Slept for 10057 us (expected: 10000 us) Offset: 57 us
2020-03-23 09:54:13,132 # Slept for 50057 us (expected: 50000 us) Offset: 57 us
2020-03-23 09:54:13,149 # Slept for 10291 us (expected: 10234 us) Offset: 57 us
2020-03-23 09:54:13,212 # Slept for 56837 us (expected: 56780 us) Offset: 57 us
2020-03-23 09:54:13,230 # Slept for 12179 us (expected: 12122 us) Offset: 57 us
2020-03-23 09:54:13,323 # Slept for 98822 us (expected: 98765 us) Offset: 57 us
2020-03-23 09:54:13,404 # Slept for 75057 us (expected: 75000 us) Offset: 57 us
2020-03-23 09:54:13,421 # Slept for 10057 us (expected: 10000 us) Offset: 57 us
2020-03-23 09:54:13,469 # Slept for 50057 us (expected: 50000 us) Offset: 57 us
2020-03-23 09:54:13,484 # Slept for 10291 us (expected: 10234 us) Offset: 57 us
2020-03-23 09:54:13,548 # Slept for 56837 us (expected: 56780 us) Offset: 57 us
2020-03-23 09:54:13,564 # Slept for 12179 us (expected: 12122 us) Offset: 57 us
2020-03-23 09:54:13,661 # Slept for 98822 us (expected: 98765 us) Offset: 57 us
2020-03-23 09:54:13,740 # Slept for 75057 us (expected: 75000 us) Offset: 57 us
2020-03-23 09:54:13,740 # Test ran for 1686119 us

```
</details>


- pr

<details><summary>BOARD=openmote-b make -C tests/xtimer_usleep flash -j3<b> offset: 58us</b></summary>

```
2020-03-23 09:50:33,143 # main(): This is RIOT! (Version: 2020.04-devel-1422-ga0f50-pr_cc2538_timer_pm)
2020-03-23 09:50:33,145 # Running test 5 times with 7 distinct sleep times
2020-03-23 09:50:33,160 # Slept for 10058 us (expected: 10000 us) Offset: 58 us
2020-03-23 09:50:33,208 # Slept for 50058 us (expected: 50000 us) Offset: 58 us
2020-03-23 09:50:33,224 # Slept for 10292 us (expected: 10234 us) Offset: 58 us
2020-03-23 09:50:33,288 # Slept for 56838 us (expected: 56780 us) Offset: 58 us
2020-03-23 09:50:33,305 # Slept for 12180 us (expected: 12122 us) Offset: 58 us
2020-03-23 09:50:33,399 # Slept for 98823 us (expected: 98765 us) Offset: 58 us
2020-03-23 09:50:33,479 # Slept for 75058 us (expected: 75000 us) Offset: 58 us
2020-03-23 09:50:33,495 # Slept for 10058 us (expected: 10000 us) Offset: 58 us
2020-03-23 09:50:33,543 # Slept for 50058 us (expected: 50000 us) Offset: 58 us
2020-03-23 09:50:33,560 # Slept for 10292 us (expected: 10234 us) Offset: 58 us
2020-03-23 09:50:33,623 # Slept for 56838 us (expected: 56780 us) Offset: 58 us
2020-03-23 09:50:33,639 # Slept for 12180 us (expected: 12122 us) Offset: 58 us
2020-03-23 09:50:33,735 # Slept for 98823 us (expected: 98765 us) Offset: 58 us
2020-03-23 09:50:33,815 # Slept for 75058 us (expected: 75000 us) Offset: 58 us
2020-03-23 09:50:33,831 # Slept for 10058 us (expected: 10000 us) Offset: 58 us
2020-03-23 09:50:33,879 # Slept for 50058 us (expected: 50000 us) Offset: 58 us
2020-03-23 09:50:33,895 # Slept for 10292 us (expected: 10234 us) Offset: 58 us
2020-03-23 09:50:33,959 # Slept for 56838 us (expected: 56780 us) Offset: 58 us
2020-03-23 09:50:33,975 # Slept for 12180 us (expected: 12122 us) Offset: 58 us
2020-03-23 09:50:34,071 # Slept for 98823 us (expected: 98765 us) Offset: 58 us
2020-03-23 09:50:34,151 # Slept for 75058 us (expected: 75000 us) Offset: 58 us
2020-03-23 09:50:34,167 # Slept for 10058 us (expected: 10000 us) Offset: 58 us
2020-03-23 09:50:34,215 # Slept for 50058 us (expected: 50000 us) Offset: 58 us
2020-03-23 09:50:34,231 # Slept for 10292 us (expected: 10234 us) Offset: 58 us
2020-03-23 09:50:34,295 # Slept for 56838 us (expected: 56780 us) Offset: 58 us
2020-03-23 09:50:34,311 # Slept for 12180 us (expected: 12122 us) Offset: 58 us
2020-03-23 09:50:34,407 # Slept for 98823 us (expected: 98765 us) Offset: 58 us
2020-03-23 09:50:34,487 # Slept for 75058 us (expected: 75000 us) Offset: 58 us
2020-03-23 09:50:34,503 # Slept for 10058 us (expected: 10000 us) Offset: 58 us
2020-03-23 09:50:34,551 # Slept for 50058 us (expected: 50000 us) Offset: 58 us
2020-03-23 09:50:34,567 # Slept for 10292 us (expected: 10234 us) Offset: 58 us
2020-03-23 09:50:34,631 # Slept for 56838 us (expected: 56780 us) Offset: 58 us
2020-03-23 09:50:34,647 # Slept for 12180 us (expected: 12122 us) Offset: 58 us
2020-03-23 09:50:34,743 # Slept for 98823 us (expected: 98765 us) Offset: 58 us
2020-03-23 09:50:34,823 # Slept for 75058 us (expected: 75000 us) Offset: 58 us
2020-03-23 09:50:34,823 # Test ran for 1686170 us

```
</details>

**run `tests/bench_xtimer`**

- master

<details><summary>BOARD=openmote-b make -C tests/bench_xtimer </summary>

```
2020-03-23 09:56:29,364 # main(): This is RIOT! (Version: 2020.04-devel-1423-gb74219-pr_cc2538_timer_pm)
2020-03-23 09:56:29,366 # xtimer benchmark application.
2020-03-23 09:56:29,366 # 
2020-03-23 09:56:29,384 #                      set() one    13160 / 1000 = 13
2020-03-23 09:56:29,385 #                   remove() one     2818 / 1000 = 2
2020-03-23 09:56:29,397 #           set() + remove() one    15686 / 1000 = 15
2020-03-23 09:56:30,245 #   set() many increasing target   842927 / 1000 = 842
2020-03-23 09:56:30,261 #                re-set()  first    13949 / 1000 = 13
2020-03-23 09:56:31,156 #                re-set() middle   888086 / 1000 = 888
2020-03-23 09:56:32,821 #                re-set()   last  1661815 / 1000 = 1661
2020-03-23 09:56:33,445 #        remove() + set()  first   614383 / 1000 = 614
2020-03-23 09:56:34,933 #        remove() + set() middle  1484569 / 1000 = 1484
2020-03-23 09:56:37,157 #        remove() + set()   last  2222876 / 1000 = 2222
2020-03-23 09:56:37,429 #       remove() many decreasing   274999 / 1000 = 274
2020-03-23 09:56:37,445 #                   xtimer_now()     5310 / 1000 = 5
2020-03-23 09:56:37,447 #               sizeof(xtimer_t)    28000 / 1000 = 28
2020-03-23 09:56:37,447 # done.
```
</details>


- pr

<details><summary>BOARD=openmote-b make -C tests/bench_xtimer </summary>

```
2020-03-23 09:56:52,387 # main(): This is RIOT! (Version: 2020.04-devel-1418-g5ef0c-HEAD)
2020-03-23 09:56:52,388 # xtimer benchmark application.
2020-03-23 09:56:52,388 # 
2020-03-23 09:56:52,404 #                      set() one    13161 / 1000 = 13
2020-03-23 09:56:52,419 #                   remove() one     2818 / 1000 = 2
2020-03-23 09:56:52,435 #           set() + remove() one    15685 / 1000 = 15
2020-03-23 09:56:53,283 #   set() many increasing target   842935 / 1000 = 842
2020-03-23 09:56:53,299 #                re-set()  first    13949 / 1000 = 13
2020-03-23 09:56:54,194 #                re-set() middle   888052 / 1000 = 888
2020-03-23 09:56:55,859 #                re-set()   last  1661756 / 1000 = 1661
2020-03-23 09:56:56,468 #        remove() + set()  first   614358 / 1000 = 614
2020-03-23 09:56:57,956 #        remove() + set() middle  1484514 / 1000 = 1484
2020-03-23 09:57:00,180 #        remove() + set()   last  2222803 / 1000 = 2222
2020-03-23 09:57:00,467 #       remove() many decreasing   274990 / 1000 = 274
2020-03-23 09:57:00,469 #                   xtimer_now()     5309 / 1000 = 5
2020-03-23 09:57:00,482 #               sizeof(xtimer_t)    28000 / 1000 = 28
2020-03-23 09:57:00,483 # done.
```
</details>